+++ src/spetlr/dbutils/__init__.py
@@ -1,1 +1,1 @@
-from .FileExists import file_exists  # noqa: F401
+from .file_exists import file_exists  # noqa: F401
